Bug 1870786 - Fix the type of RTP timestamps in WebrtcGmpVideoEncoder. r=webrtc-reviewers,pehrsons,media-playback-reviewers,aosmond
The keys inserted in the map come from GMPVideoEncodedFrame::TimeStamp,
which returns a uint64_t, and truncation is probably not wanted,
although it might just work in practice.

Differential Revision: https://phabricator.services.mozilla.com/D197012

+++ dom/media/gtest/TestGMPRemoveAndDelete.cpp
@@ -361,7 +361,7 @@ void GMPRemoveTest::gmp_Decode() {
       uint8_t y_;
       uint8_t u_;
       uint8_t v_;
-      uint32_t timestamp_;
+      uint64_t timestamp_;
     } idr_nalu;
   };
 #pragma pack(pop)
